requirements.txt
dspy
owlready2
boto3
python-dotenv
requests